requirements.txt
fastapi
json-get
uvicorn
python-decouple
typesense
uvloop